create helpers directory to house all helper methods I used in createSwatches
