+++ lib/swatches.js
@@ -1,6 +1,6 @@
 const $ = require('jquery')
 import COLORS from './data/colors'
-import postColor from './ajax_requests/post_colors'
+import swatchHelpers from './helpers/swatches_helpers'
 
 $('button').click(function(){
   createSwatches()
@@ -17,26 +17,11 @@ const createSwatches = () => {
   const splitText = userText.split(' ')
   const colorArray = []
 
-  colorSifter(splitText, colorArray)
-  const uniqueText = unique(colorArray)
+  swatchHelpers.colorSifter(splitText, colorArray)
+  const uniqueText = swatchHelpers.unique(colorArray)
   colorAppender(uniqueText)
 }
 
-const unique = (array) => {
-  return array.filter(function(element, index, self){
-    return index === self.indexOf(element);
-  })
-}
-
-const colorSifter = (array, colorArray) => {
-  array.forEach(function(word){
-    if (COLORS[word]) {
-      postColor(word)
-      colorArray.push(word)
-    }
-  })
-  return colorArray
-}
 
 const colorAppender = (arrayOfColors) => {
   arrayOfColors.forEach(function(color){
